cancel adding donation modal
