+++ client/src/components/donations/addDonations/donationsCard.tsx
@@ -1,77 +1,13 @@
 import React, {useState} from 'react'
 import { Select } from '@chakra-ui/react'
-import { Card, CardHeader, CardBody, CardFooter, Text,  Input, InputGroup, InputRightAddon, InputLeftAddon} from '@chakra-ui/react'
+import { Card, CardHeader, CardBody, CardFooter, Text } from '@chakra-ui/react'
 
+import DonationInputs, { DonationSub } from "./donationInputs";
 export interface Donation {
     donor: string,
     sub: DonationSub[];
 }
 
-interface DonationSub {
-    date: Date, 
-    category: string, 
-    weight: number,
-    value: number,
-}
-
-function DonationInputs({ subDonation, index, onChange }: { 
-    subDonation: DonationSub; 
-    index: number; 
-    onChange: (index: number, e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>) => void;
-  }) {
-      return (
-          <CardBody display="flex" >
-            <CardBody>
-              <Text>Date Donated</Text>
-              <Input
-                  name="date"
-                  type="date"
-                  value={subDonation.date instanceof Date ? subDonation.date.toISOString().split('T')[0] : ''}
-                  onChange={(e) => onChange(index, e)}
-              />
-            </CardBody>
-              
-            <CardBody>
-              <Text>Category</Text>
-              <Select
-                  name="category"
-                  value={subDonation.category}
-                  placeholder="Select Category"
-                  onChange={(e) => onChange(index, e)}
-              >
-                  <option value="food">Food</option>
-                  <option value="client">Client</option>
-              </Select>
-            </CardBody>
-            <CardBody>
-              <Text>Total Weight</Text>
-              <InputGroup>
-                  <Input
-                      name="weight"
-                      type="number"
-                      placeholder="XX"
-                      value={isNaN(subDonation.weight) ? "" : subDonation.weight}
-                      onChange={(e) => onChange(index, e)}
-                  />
-                  <InputRightAddon children="lbs" />
-              </InputGroup>
-            </CardBody>
-            <CardBody>
-              <Text>Value per lbs</Text>
-              <InputGroup>
-                  <InputLeftAddon children="$" />
-                  <Input
-                      name="value"
-                      type="number"
-                      placeholder="0.00"
-                      value={isNaN(subDonation.value) ? "" : subDonation.value}
-                      onChange={(e) => onChange(index, e)}
-                  />
-              </InputGroup>
-              </CardBody>
-          </CardBody>
-      );
-  }
 function DonationCard({ donationToEdit, onSubmit }: { donationToEdit?: Donation, onSubmit: (donation: Donation) => void }) {
     const [donation, setDonation] = useState<Donation>(
       donationToEdit || {
